fixed regex matching and Entity deletion

+++ backend/tomato/host/site.py
@@ -53,7 +53,8 @@ class Site(Entity, BaseDocument):
 
 	def _remove(self):
 		logging.logMessage("remove", category="site", name=self.name)
-		self.delete()
+		if self.id:
+			self.delete()
 
 	ACTIONS = {
 		Entity.REMOVE_ACTION: Action(fn=_remove, check=_checkRemove)
@@ -103,7 +104,7 @@ class Site(Entity, BaseDocument):
 		logging.logMessage("create", category="site", name=name, label=label)
 		site = Site(name=name, organization=orga, label=label)
 		try:
-			attrs_ = {k: v for k, v in attrs.iteritems()}
+			attrs_ = attrs.copy()
 			attrs_['name'] = name
 			attrs_['label'] = label
 			site.init(attrs_)
